added Drawing class as helper
